Implement a separate vagrant connection type to channel SSH through the vagrant mechanism to deal with managed key pairs and transaparent port forwarding.
Can be used via makegateway("vagrant=<machine>") and accepts all additional
SSH options as SSH does.

The test is a bit annoying as it currently demonstrates this working by
actually firing up vagrant. This means it a) needs vagrant and virtualbox and
b) it downloads a relatively large image when running the tests the first time
and c) it takes quite a few seconds running the test at all times.